move phantom block logic into consensus graph
