+++ client/src/rpc/types/eth/mod.rs
@@ -12,7 +12,7 @@ mod sync;
 mod transaction;
 
 pub use self::{
-    block::{Block, PhantomBlock},
+    block::Block,
     block_number::BlockNumber,
     call_request::CallRequest,
     filter::{EthRpcLogFilter, FilterChanges},
