+++ core/src/consensus/mod.rs
@@ -9,15 +9,14 @@ pub mod debug_recompute;
 mod pastset_cache;
 pub mod pos_handler;
 
-pub use crate::consensus::{
-    consensus_inner::{ConsensusGraphInner, ConsensusInnerConfig},
-    consensus_trait::{ConsensusGraphTrait, SharedConsensusGraph},
-};
-
 use super::consensus::consensus_inner::{
     confirmation_meter::ConfirmationMeter,
     consensus_executor::ConsensusExecutor,
     consensus_new_block_handler::ConsensusNewBlockHandler,
+};
+pub use crate::consensus::{
+    consensus_inner::{ConsensusGraphInner, ConsensusInnerConfig},
+    consensus_trait::{ConsensusGraphTrait, SharedConsensusGraph},
 };
 use crate::{
     block_data_manager::{
@@ -29,7 +28,9 @@ use crate::{
         },
         pos_handler::PosVerifier,
     },
-    executive::ExecutionOutcome,
+    executive::{
+        internal_contract::build_bloom_and_recover_phantom, ExecutionOutcome,
+    },
     observer::{
         trace::{ActionType, BlockExecTraces, LocalizedTrace},
         trace_filter::TraceFilter,
@@ -51,7 +52,7 @@ use cfx_parameters::{
 use cfx_state::state_trait::StateOpsTrait;
 use cfx_statedb::StateDb;
 use cfx_storage::state_manager::StateManagerTrait;
-use cfx_types::{AddressWithSpace, AllChainID, Bloom, H256, U256};
+use cfx_types::{AddressWithSpace, AllChainID, Bloom, Space, H256, U256};
 use either::Either;
 use itertools::Itertools;
 use malloc_size_of::{MallocSizeOf, MallocSizeOfOps};
@@ -67,7 +68,8 @@ use primitives::{
     log_entry::LocalizedLogEntry,
     pos::PosBlockId,
     receipt::Receipt,
-    EpochId, EpochNumber, SignedTransaction, TransactionIndex,
+    EpochId, EpochNumber, PhantomBlock, SignedTransaction, TransactionIndex,
+    TransactionOutcome,
 };
 use rayon::prelude::*;
 use std::{
@@ -1487,6 +1489,132 @@ impl ConsensusGraph {
             }
         }
         Ok(traces)
+    }
+
+    pub fn get_phantom_block_by_number(
+        &self, block_num: EpochNumber, pivot_assumption: Option<H256>,
+    ) -> Result<Option<PhantomBlock>, String> {
+        let hashes = self.get_block_hashes_by_epoch(block_num)?;
+
+        let blocks = match self
+            .get_data_manager()
+            .blocks_by_hash_list(&hashes, false /* update_cache */)
+        {
+            None => return Ok(None),
+            Some(b) => b,
+        };
+
+        // sanity check: epoch is not empty
+        let pivot = match blocks.last() {
+            Some(p) => p,
+            None => return Err("Inconsistent state".into()),
+        };
+
+        if matches!(pivot_assumption, Some(h) if h != pivot.hash()) {
+            return Ok(None);
+        }
+
+        let mut phantom_block = PhantomBlock {
+            pivot_header: pivot.block_header.clone(),
+            transactions: vec![],
+            receipts: vec![],
+            errors: vec![],
+        };
+
+        let mut gas_used = U256::from(0);
+
+        for b in &blocks {
+            // note: we need the receipts to reconstruct a phantom block.
+            // as a result, we cannot return unexecuted blocks in eth_* RPCs.
+            let exec_info = match self
+                .get_data_manager()
+                .block_execution_result_by_hash_with_epoch(
+                    &b.hash(),
+                    &pivot.hash(),
+                    false, // update_pivot_assumption
+                    false, // update_cache
+                ) {
+                None => return Ok(None),
+                Some(r) => r,
+            };
+
+            let block_receipts = &exec_info.block_receipts.receipts;
+            let errors = &exec_info.block_receipts.tx_execution_error_messages;
+
+            // sanity check: transaction and
+            if b.transactions.len() != block_receipts.len() {
+                return Err("Inconsistent state".into());
+            }
+
+            let evm_chain_id = self.best_chain_id().in_evm_space();
+
+            for (id, tx) in b.transactions.iter().enumerate() {
+                match tx.space() {
+                    Space::Ethereum => {
+                        let receipt = &block_receipts[id];
+
+                        // we do not return non-executed transaction
+                        if receipt.outcome_status == TransactionOutcome::Skipped
+                        {
+                            continue;
+                        }
+
+                        phantom_block.transactions.push(tx.clone());
+
+                        // sanity check: gas price must be positive
+                        if *tx.gas_price() == 0.into() {
+                            return Err("Inconsistent state".into());
+                        }
+
+                        // FIXME(thegaram): is this correct?
+                        gas_used += receipt.gas_fee / tx.gas_price();
+
+                        phantom_block.receipts.push(Receipt {
+                            accumulated_gas_used: gas_used,
+                            outcome_status: receipt.outcome_status,
+                            ..receipt.clone()
+                        });
+
+                        phantom_block.errors.push(errors[id].clone());
+                    }
+                    Space::Native => {
+                        let (phantom_txs, _) = build_bloom_and_recover_phantom(
+                            &block_receipts[id].logs[..],
+                            tx.hash(),
+                        );
+
+                        for p in phantom_txs {
+                            phantom_block.transactions.push(Arc::new(
+                                p.clone().into_eip155(evm_chain_id),
+                            ));
+
+                            // note: phantom txs consume no gas
+                            let phantom_receipt = p.into_receipt(gas_used);
+                            phantom_block.receipts.push(phantom_receipt);
+
+                            // FIXME(thegaram): handle errors for phantom txs
+                            phantom_block.errors.push("".into());
+                        }
+                    }
+                }
+            }
+        }
+
+        Ok(Some(phantom_block))
+    }
+
+    pub fn get_phantom_block_by_hash(
+        &self, hash: &H256,
+    ) -> Result<Option<PhantomBlock>, String> {
+        let epoch_num = match self.get_block_epoch_number(hash) {
+            None => return Ok(None),
+            Some(n) => n,
+        };
+
+        self.get_phantom_block_by_number(
+            EpochNumber::Number(epoch_num),
+            Some(*hash),
+        )
     }
 }
